app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Marsel <lllmarselll@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.3"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
flake8 = "^7.1.1"
mypy = "^1.12.1"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
asyncpg = "^0.30.0"
python-dotenv = "^1.0.1"
redis = "^5.2.0"
setuptools = "^75.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
